raiting/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "tortoise-orm>=0.25.0",
    "uvicorn>=0.34.2",
]
